feat: implement basic functionalities
This commit introduces the field comparison function. To achieve similar
UX, two fragments are used, both of which are sourced from Android
Studio's template. The following flow is achieved through this commit:

1. The user enters texts into the two EditText
2. After the check button is pressed, compare the two texts
3. With the result provided, pass a message to the next fragment; this
is achieved by using the MainActivity as an intermediary. The message is
in a form of a boolean, "SAME" or "NOT THE SAME"
4. The user is shown the message, displayed through the EditText in the
second fragment

There are far better ways to do this, however, this contains fewer
complexity compared to other methods, as it relies only on several basic
calls and patterns. Future commits should handle the hack-job done here,
if necessary.

PS1: Remove boilerplates.